feat(theme): migrate dashboard into (dj) route group, retire events list page
Move app/dashboard/page.tsx → app/(dj)/dashboard/page.tsx so the page
sits inside the (dj) layout and inherits ThemeToggle. Merge in HelpSpot/
HelpButton/OnboardingOverlay from the unreachable events list page. Delete
the old dashboard and events pages; move ActivityLogPanel to the new
co-located components dir. 25 new tests cover all behaviour.
