+++ dashboard/app/(dj)/dashboard/components/ActivityLogPanel.tsx
